requirements.txt
future
emoji
requests
sqlalchemy
python-telegram-bot==11.1.0
psycopg2-binary
feedparser
pyyaml
pyowm==2.10.0
